requirements.txt
# LLM 관련
transformers>=4.35.2
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  # 4bit/8bit 양자화 지원

# 파인튜닝 관련
peft>=0.7.0  # QLoRA/LoRA
trl>=0.7.0  # SFTTrainer
datasets>=2.14.0  # 데이터셋 로딩
tqdm>=4.66.0  # 진행바

# 모델 서버
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# HTTP 클라이언트
requests>=2.31.0
openai>=1.0.0

# 데이터베이스 연동
pymysql>=1.1.0
sqlalchemy>=2.0.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 유틸리티
python-dotenv>=1.0.0
pyyaml>=6.0
